+++ src/components/App.styles.ts
@@ -1,0 +1,14 @@
+import { createUseStyles } from "react-jss";
+
+export default createUseStyles({
+  root: {},
+  recordContainer: {
+    alignSelf: "center",
+    alignItems: "center",
+    height: "100%",
+    flexDirection: "column",
+    margin: "1rem",
+    justifyContent: "space-between",
+    minWidth: "70%",
+  },
+});