+++ src/components/MainContent/MainContent.styles.ts
@@ -1,7 +1,8 @@
 import { createUseStyles } from "react-jss";
 
 export default createUseStyles({
-  root: {
-    padding: "0.5rem",
-  },
+  root: (props) => ({
+    margin: props.isMobile ? "1rem" : 0,
+    width: props.isMobile ? window.innerWidth * 0.9 : 320,
+  }),
 });