+++ src/components/MainContent/MainContent.tsx
@@ -3,13 +3,16 @@ import React from "react";
 import Scores from "components/Scores";
 import useStyles from "./MainContent.styles";
 
-export default function MainContent(): JSX.Element {
-  const classes = useStyles();
+export default function MainContent({
+  isMobile,
+}: {
+  isMobile: boolean;
+}): JSX.Element {
+  const classes = useStyles({ isMobile });
 
   return (
     <div id="main" className={classes.root}>
-      <Scores />
-
+      {!isMobile && <Scores />}
       <PlayersSection />
     </div>
   );