+++ src/components/MapsContent/MapsContent.tsx
@@ -1,12 +1,20 @@
+import Button from "components/common/Button";
 import Draggable from "react-draggable";
 import MiraHq from "./MiraHq";
 import Polus from "./Polus";
 import React from "react";
 import TheSkeld from "./TheSkeld";
 import useStyles from "./MapsContent.styles";
-import Button from "components/common/Button";
 
-export default function MapsContent(): JSX.Element {
+// export interface IMapsContentProps {
+//   isMobile: boolean;
+// }props: IMapsContentProps
+
+export default function MapsContent({
+  isMobile,
+}: {
+  isMobile: boolean;
+}): JSX.Element {
   const [map, setMap] = React.useState("skeld");
   const [resetState, setResetState] = React.useState(false);
 
@@ -27,6 +35,7 @@ export default function MapsContent(): JSX.Element {
 
   const classes = useStyles({
     map: map === "skeld" ? "TheSkeld" : map === "mira" ? "Mirahq" : "Polus",
+    isMobile,
   });
 
   let currentMap = <TheSkeld />;
@@ -70,23 +79,29 @@ export default function MapsContent(): JSX.Element {
         />
       </div>
       <div className={classes.wrapper}>
-        <Draggable bounds="parent" position={{ x: 0, y: 0 }} disabled>
-          {currentMap}
-        </Draggable>
+        {currentMap}
+
         {!resetState &&
           players.map((player) => (
             <Draggable key={player} bounds="parent">
               <img
                 src={`assets/${player}.png`}
                 className={classes.playerIcon}
+                onDrag={(event: React.DragEvent<HTMLImageElement>) =>
+                  event.stopPropagation()
+                }
+                draggable={false}
               />
             </Draggable>
           ))}
-
-        <Button className={classes.button} onClick={() => setResetState(true)}>
-          Reset Players
-        </Button>
       </div>
+      <Button className={classes.button} onClick={() => setResetState(true)}>
+        Reset Players
+      </Button>
     </div>
   );
 }
+
+MapsContent.defaultProps = {
+  isMobile: false,
+};