+++ src/components/Player/Player.styles.ts
@@ -12,8 +12,16 @@ export default createUseStyles((theme: ITheme) => ({
     border: `1px solid ${theme.borderColor}`,
     borderRadius: "0.25rem",
     display: "flex",
+    flex: `0 1 ${
+      props.isMobile && window.screen.orientation.type.includes("portrait")
+        ? "100%"
+        : props.names
+        ? "50%"
+        : "25%"
+    }`,
     justifyContent: "center",
     margin: "0.25rem 0",
+    opacity: props.longPressed ? "0.5" : 1,
     padding: "0.25rem 0.1rem",
     position: "relative",
   }),
@@ -26,7 +34,7 @@ export default createUseStyles((theme: ITheme) => ({
   icon: (props) => ({
     borderRadius: "6px",
     padding: "0.15rem",
-    margin: "0 0.15rem",
+    margin: props.isMobile ? "0 0.5rem" : "0 0.15rem",
     display: "flex",
     justifyContent: "center",
 
@@ -43,7 +51,7 @@ export default createUseStyles((theme: ITheme) => ({
   }),
   input: (props) => ({
     borderRadius: ".2rem",
-    fontSize: props.name ? "1.15rem" : "1rem",
+    fontSize: props.isMobile ? "1.5rem" : props.name ? "1.15rem" : "1rem",
     fontWeight: 700,
     letterSpacing: "0.025rem",
     lineHeight: 1.5,