+++ src/components/Player/Player.tsx
@@ -1,27 +1,34 @@
 import ColorsMenu from "../ColorsMenu";
 import { IPlayer } from "utils/types";
-import Input from "components/common/Input";
 import React from "react";
 import { useData } from "context";
-import useStyles from "./Player.styles";
-import { useTranslation } from "react-i18next";
+import { MobileContext } from "components/App";
+import usePlayerStyles from "./Player.styles";
 
 export interface IPlayerProps {
   id: string | number;
   color: string;
+  backgroundColor: string;
   name: string;
   list: Array<IPlayer>;
-  listName: string;
   setList: (value: IPlayer[]) => void;
   index: number;
 }
 
 export default function Player(props: IPlayerProps): JSX.Element {
   const [isMenuShowing, setIsMenuShowing] = React.useState(false);
+  const [longPressed, setLongPressed] = React.useState(false);
+  const isMobile = React.useContext(MobileContext);
 
   const { names } = useData()!; // eslint-disable-line
-  const { t } = useTranslation();
-  const classes = useStyles({ names, ...props });
+  const htmlElRef = React.useRef(null);
+
+  const playerStyles = usePlayerStyles({
+    names,
+    isMobile,
+    longPressed,
+    ...props,
+  });
 
   const { id, color, name, list, setList, index } = props;
 
@@ -30,14 +37,57 @@ export default function Player(props: IPlayerProps): JSX.Element {
     event: React.ChangeEvent<HTMLInputElement>
   ) => {
     const players: Array<IPlayer> = [...list];
-
     players[player].name = event.currentTarget.value;
-
     setList(players);
   };
 
+  const handleKeyPress = (event: React.KeyboardEvent<HTMLInputElement>) => {
+    if (event.key === "Enter") {
+      const currentInput = (htmlElRef.current as unknown) as HTMLInputElement;
+      const nextParent =
+        currentInput.parentElement?.parentElement?.nextElementSibling ??
+        currentInput.parentElement?.parentElement?.parentElement
+          ?.firstElementChild;
+      const nextInput = nextParent?.lastChild?.firstChild as HTMLInputElement;
+      nextInput?.select();
+    }
+  };
+
+  // put this in for mobile only
+  // so users know they are interacting with the object
+  const useLongPress = (ms = 150) => {
+    const [startLongPress, setStartLongPress] = React.useState(false);
+
+    React.useEffect(() => {
+      let timerId: number;
+      if (startLongPress) {
+        timerId = setTimeout(() => setLongPressed(true), ms);
+      } else {
+        setLongPressed(false);
+        clearTimeout(timerId);
+      }
+
+      return () => {
+        clearTimeout(timerId);
+      };
+    }, [ms, startLongPress]);
+
+    return {
+      onMouseDown: () => setStartLongPress(true),
+      onMouseUp: () => setStartLongPress(false),
+      onMouseLeave: () => setStartLongPress(false),
+      onTouchStart: () => setStartLongPress(true),
+      onTouchEnd: () => setStartLongPress(false),
+    };
+  };
+
+  const longPressEvents = useLongPress();
+
   return (
-    <div className={`${classes.container} player-handle`}>
+    <div
+      className={`${playerStyles.container} player-handle`}
+      {...longPressEvents}
+    >
       {isMenuShowing && (
         <ColorsMenu
           isMenuShowing={isMenuShowing}
@@ -45,7 +95,7 @@ export default function Player(props: IPlayerProps): JSX.Element {
           currentColor={id}
         />
       )}
-      <div className={classes.icon}>
+      <div className={playerStyles.icon}>
         <img
           onClick={() => {
             if (names) {
@@ -58,14 +108,17 @@ export default function Player(props: IPlayerProps): JSX.Element {
         />
       </div>
       {names && (
-        <div className={classes.name}>
-          <Input
-            placeholder={t("main.player")}
-            className={classes.input}
+        <div className={playerStyles.name}>
+          <input
+            type="text"
+            placeholder="Player"
+            className={playerStyles.input}
             onChange={(event: React.ChangeEvent<HTMLInputElement>) =>
               handleChange(index, event)
             }
+            onKeyPress={handleKeyPress}
             value={name}
+            ref={htmlElRef}
           />
         </div>
       )}