+++ src/components/PlayersSection/PlayersSection.styles.ts
@@ -1,5 +1,5 @@
 import { createUseStyles } from "react-jss";
 
 export default createUseStyles({
-  root: { padding: "0.1rem" },
+  root: { padding: "0.1rem", width: "100%" },
 });