+++ src/components/Scores/Scores.tsx
@@ -4,11 +4,9 @@ import React from "react";
 import { useData } from "context";
 import useStyles from "./Scores.styles";
 import { useTheme } from "react-jss";
-import { useTranslation } from "react-i18next";
 
 export default function Scores(): JSX.Element {
   const theme = useTheme<ITheme>();
-  const { t } = useTranslation();
   const classes = useStyles();
 
   const {
@@ -32,53 +30,44 @@ export default function Scores(): JSX.Element {
   const impostorRate = getRate(impostorWins, impostorGames);
 
   return (
-    <React.Fragment>
-      <div className={classes.root}>
-        <div className={classes.title}>
-          <span>{t("main.overall")}</span>
-          <span>
-            {overallWins}
-            {t("main.w")} - {overallLosses}
-            {t("main.l")}
-          </span>
-        </div>
+    <div className={classes.root}>
+      <div className={classes.progressBarContainer}>
+        <span className={classes.title}>Overall</span>
+
         <ProgressBar
           progress={overallRate}
           backgroundColor={theme.neutralBackgroundColor}
           progressColor={theme.neutralTextColor}
           className={classes.progress}
+          wins={overallWins}
+          losses={overallLosses}
         />
+      </div>
 
-        <div className={classes.title}>
-          <span>{t("main.innocent")}</span>
-          <span>
-            {innocentWins}
-            {t("main.w")} - {innocentLosses}
-            {t("main.l")}
-          </span>
-        </div>
+      <div className={classes.progressBarContainer}>
+        <span className={classes.title}>Innocent</span>
         <ProgressBar
           progress={innocentRate}
           backgroundColor={theme.innocentBackgroundColor}
           progressColor={theme.innocentTextColor}
           className={classes.progress}
+          wins={innocentWins}
+          losses={innocentLosses}
         />
+      </div>
 
-        <div className={classes.title}>
-          <span>{t("main.impostor")}</span>
-          <span>
-            {impostorWins}
-            {t("main.w")} - {impostorLosses}
-            {t("main.l")}
-          </span>
-        </div>
+      <div className={classes.progressBarContainer}>
+        <span className={classes.title}>Impostor</span>
+
         <ProgressBar
           progress={impostorRate}
           backgroundColor={theme.impostorBackgroundColor}
           progressColor={theme.impostorTextColor}
           className={classes.progress}
+          wins={impostorWins}
+          losses={impostorLosses}
         />
       </div>
-    </React.Fragment>
+    </div>
   );
 }