+++ src/components/Section/Section.tsx
@@ -9,15 +9,16 @@ import useStyles from "./Section.styles";
 export interface ISectionProps {
   title: string;
   list: Array<IPlayer>;
+  listName: string;
   setList: (value: IPlayer[]) => void;
+  isMobile: boolean;
 }
 
 export default function Section(props: ISectionProps): JSX.Element {
   const { names } = useData()!; // eslint-disable-line
   const classes = useStyles({ names });
 
-  const { title, list, setList } = props;
-
+  const { isMobile, title, list, setList } = props;
   return (
     <React.Fragment>
       <div className={classes.root}>
@@ -26,6 +27,8 @@ export default function Section(props: ISectionProps): JSX.Element {
         <ReactSortable
           group="players"
           handle=".player-handle"
+          delayOnTouchStart={isMobile}
+          delay={isMobile ? 150 : 0}
           list={list}
           setList={setList}
           className={classes.players}