+++ src/components/common/Switch/Switch.styles.ts
@@ -8,8 +8,8 @@ export default createUseStyles((theme: ITheme) => ({
     "-webkit-appearance": "none",
     "-moz-appearance": "none",
     appearance: "none",
-    width: "3rem",
-    height: "1rem",
+    minWidth: "3rem",
+    height: "1.1rem",
     display: "inline-block",
     position: "relative",
     borderRadius: "2rem",
@@ -19,7 +19,6 @@ export default createUseStyles((theme: ITheme) => ({
     cursor: "pointer",
     backgroundColor: `rgba(${hexToRGB(theme.buttonBackgroundColor)}, 0.5)`,
     transition: "background-color ease 0.3s",
-    margin: "auto",
     marginRight: "0.5rem",
 
     "&:before": {
@@ -47,5 +46,12 @@ export default createUseStyles((theme: ITheme) => ({
     "&:checked:before": {
       left: "2rem",
     },
+  },
+  toggleContainer: {
+    display: "flex",
+    alignItems: "center",
+    fontSize: "1rem",
+    justifyContent: "center",
+    margin: "0.5rem 0",
   },
 }));